requirements.txt
Django~=4.0
Pillow==9.2.0
django-admin-sortable2==2.1